simplify: shared modules for notifications, serializer helpers, and match errors
/simplify cleanup of the MCP-over-shared-services branch (behavior-preserving):
- Relocate the result-notification cluster (notify_result_posted + copy helpers)
  out of the matches router into app/match_result_notifications.py, so
  app/mcp_server.py is genuinely router-free (no app.matches import).
- Dedup the verbatim _view_extras and read-path _load_match into
  app/match_serialization.py (public view_extras / load_match_eager).
- Move the whole match-flow domain-exception family into a neutral
  app/match_errors.py; result_acceptance re-exports it. (accept_result keeps its
  behavior-identical lazy load_match default — the real import cycle is driven by
  match_queries importing scoring primitives from result_acceptance, out of scope.)
- Drop two dead ToolError branches in _map_score_write_tool_error; reuse the
  existing _game_by_number helper at four sites.

No route/schema/docstring change; OpenAPI identical; full suite 1449 passed.
