+++ api/app/matches.py
@@ -25,23 +25,32 @@ from app.attention import (
     list_attention_kind,
 )
 from app.db import get_session
-from app.mappers.match_extras_mapper import (
-    MatchDetailsExtras,
-    empty_extras,
-    serialize_match_extras,
-)
+from app.mappers.match_extras_mapper import empty_extras
 from app.match_creation import create_match as create_match_core
+from app.match_errors import (
+    CannotAcceptOwnProposalError,
+    MatchClosedError,
+    MatchNotFoundError,
+    MatchNotScorableError,
+    NegotiationConflictError,
+    OpponentNotFoundError,
+    PostedGamesNotDecisiveError,
+    RatedNeedsRegisteredOpponentError,
+    ResultNotFoundError,
+    ScoreConflictError,
+    ScoreNotAllowedError,
+    SelfMatchError,
+    UndecidedBoardError,
+)
 from app.match_queries import (
     _actionable_attention_filter,
     _attention_matches_query,
     _player_username_filter,
     current_game_number,
     match_eager_options,
-    my_side,
-    opponent_side,
     participant_filter,
-    singles_user_ids,
 )
+from app.match_result_notifications import notify_result_posted
 from app.match_scoring import (
     MatchLockUnavailable,
     enter_game_score,
@@ -66,6 +75,8 @@ from app.match_serialization import (
     _serialize_details,
     _side_schema,
     _status_label,
+    load_match_eager,
+    view_extras,
 )
 from app.models import (
     Match,
@@ -73,29 +84,13 @@ from app.models import (
     MatchStatus,
     User,
 )
-from app.notifications.apns import MATCH_RESULT_CONFIRMATION_CATEGORY
 from app.notifications.dependencies import get_notification_service
 from app.notifications.service import NotificationService
-from app.notifications.taxonomy import NotificationCategory
 from app.rate_limiting import RedisRateLimiter
 from app.result_acceptance import (
-    CannotAcceptOwnProposalError,
-    MatchClosedError,
-    MatchNotFoundError,
-    MatchNotScorableError,
-    NegotiationConflictError,
-    OpponentNotFoundError,
-    PostedGamesNotDecisiveError,
-    RatedNeedsRegisteredOpponentError,
-    ResultNotFoundError,
-    ScoreConflictError,
-    ScoreNotAllowedError,
-    SelfMatchError,
-    UndecidedBoardError,
     accept_result,
     side_win_counts,
 )
-from app.result_chain import standing_result
 
 # Re-exported for ``tests/test_ratings.py``, which loads the finalize eager-load
 # superset via ``app.matches.match_rating_eager_options`` (the propose/accept
@@ -119,7 +114,6 @@ from app.schemas.match import (
     MatchListRow,
     MatchResultsWrite,
 )
-from app.schemas.notification import NotificationJob
 from app.services.dependencies import get_match_service
 from app.services.match_service import MatchService
 from app.sessions import get_current_user, get_optional_user
@@ -132,20 +126,6 @@ MAX_PAGE_SIZE = 100
 
 
 # ----- helpers -------------------------------------------------------------
-
-
-async def _load_match(
-    db: AsyncSession,
-    match_id: uuid.UUID,
-    *,
-    options: tuple[ExecutableOption, ...] | None = None,
-) -> Match | None:
-    result = await db.execute(
-        select(Match)
-        .where(Match.id == match_id)
-        .options(*(options if options is not None else match_eager_options()))
-    )
-    return result.scalar_one_or_none()
 
 
 # One message for every score-write conflict — a concurrent participant already
@@ -168,25 +148,6 @@ def _score_conflict(committed: MatchDetailsScore | None) -> HTTPException:
             message=_SCORE_CONFLICT_MESSAGE,
             committed_score=committed,
         ).model_dump(mode="json"),
-    )
-
-
-async def _view_extras(match_service: MatchService, match: Match) -> MatchDetailsExtras:
-    """The participant-only extras block (rating changes, recent form,
-    head-to-head) for an already-loaded ``match``.
-
-    The router hands the service the primitives it already holds and gets back a
-    domain model; the SQL lives in ``MatchDetailsRepository`` and the wire shapes
-    are built by the extras mapper. Callers gate on participation *before* calling
-    this — a non-participant gets ``empty_extras()`` (see #515)."""
-    return serialize_match_extras(
-        await match_service.load_view_extras(
-            match_id=match.id,
-            league_id=match.league_id,
-            status=match.status,
-            created_at=match.created_at,
-            user_ids=singles_user_ids(match),
-        )
     )
 
 
@@ -567,7 +528,7 @@ async def get_match(
 
     The serializer flags whether the current user is on a side; write paths
     below still gate on participation via `get_current_user`."""
-    match = await _load_match(db, match_id)
+    match = await load_match_eager(db, match_id)
     if match is None:
         raise HTTPException(status_code=404, detail="Match not found.")
     domain_match = await match_service.get_match(match_id)
@@ -580,7 +541,7 @@ async def get_match(
         match, current_user.id
     )
     extras = (
-        await _view_extras(match_service, match) if is_participant else empty_extras()
+        await view_extras(match_service, match) if is_participant else empty_extras()
     )
     return _serialize_details(
         match,
@@ -588,130 +549,6 @@ async def get_match(
         extras,
         domain_match,
     )
-
-
-# ----- result-acceptance push ---------------------------------------------
-
-# En-dash between scores reads better than a hyphen in notification copy.
-_SCORE_DASH = "–"
-
-
-def _game_scores_text(match: Match, poster_side_number: int) -> str:
-    """Compact per-game scores oriented poster-first (e.g. ``"11–7, 9–11"``),
-    so they line up with the games-won headline the recipient sees. Games with
-    no saved score are skipped."""
-    parts: list[str] = []
-    for game in sorted(match.games, key=lambda g: g.game_number):
-        if game.score is None:
-            continue
-        if poster_side_number == 1:
-            poster_pts, recipient_pts = (
-                game.score.side_1_points,
-                game.score.side_2_points,
-            )
-        else:
-            poster_pts, recipient_pts = (
-                game.score.side_2_points,
-                game.score.side_1_points,
-            )
-        parts.append(f"{poster_pts}{_SCORE_DASH}{recipient_pts}")
-    return ", ".join(parts)
-
-
-def _result_confirmation_copy(
-    match: Match, poster_id: uuid.UUID, *, is_counter: bool
-) -> tuple[str, str] | None:
-    """Title + body for the "accept or suggest a correction to the result your
-    opponent posted" push, framed for the *recipient* (the side that didn't
-    post).
-
-    ``is_counter`` picks the closing prompt to match the actual button pair
-    the recipient's in-app callout renders: a first-post shows Accept /
-    Suggest correction (``confirmation-callout-display.tsx``'s ``"review"``
-    case), while a counter shows Accept / Counter (its ``"corrected"`` case,
-    #728). The native iOS push notification itself only ever offers a single,
-    static Approve/Suggest-correction action pair (``PushNotificationManager
-    .swift``) — it doesn't yet grow a counter-specific action — so this body
-    text is the only place a tapped-through counter reads "Counter" until the
-    push actions themselves are split the same way.
-
-    The headline carries the games-won score and, where there's room, the
-    body lists the individual game scores — both oriented so the poster's
-    number comes first. Returns ``None`` when the match isn't a two-human
-    match (nothing to accept)."""
-    poster_side = my_side(match, poster_id)
-    recipient_side = opponent_side(match, poster_id)
-    if poster_side is None or recipient_side is None or not poster_side.players:
-        return None
-
-    poster_name = poster_side.players[0].user.username
-    wins = side_win_counts(match)
-    poster_games = wins.get(poster_side.side_number, 0)
-    recipient_games = wins.get(recipient_side.side_number, 0)
-
-    # Score always reads winner-first; the phrase tells the recipient which
-    # side the poster claims won.
-    if poster_games >= recipient_games:
-        phrase, hi, lo = "beating you", poster_games, recipient_games
-    else:
-        phrase, hi, lo = "losing to you", recipient_games, poster_games
-    headline = f"{poster_name} reported {phrase} {hi}{_SCORE_DASH}{lo}"
-
-    prompt = "Accept or counter?" if is_counter else "Accept or suggest a correction?"
-    games = _game_scores_text(match, poster_side.side_number)
-    body = f"{headline}. Games: {games}. {prompt}" if games else f"{headline}. {prompt}"
-    return "Accept your match result", body
-
-
-async def _notify_result_posted(
-    notifications: NotificationService, match: Match, poster_id: uuid.UUID
-) -> None:
-    """Queue an accept/counter (or accept/suggest-correction) prompt to every
-    player on the side that now owes a response. Each enqueued job persists
-    the in-app record (the bell feed) and fans out push/email per the
-    recipient's preferences in the worker. The APNs ``category``/``data``
-    carry the action group, the match id (so a tapped push deep-links to the
-    right match), and the standing result id (so a tapped Approve binds to the
-    exact result the push described, not whatever is standing at tap time — see
-    docs/adr/0007), plus a per-match ``collapse_id`` so a superseding push
-    replaces the stale one on the lock screen."""
-    recipient_side = opponent_side(match, poster_id)
-    if recipient_side is None or not recipient_side.players:
-        return
-    # Derive counter-vs-first-post from the same viewer-relative negotiation
-    # state the BFF/UI use (``_negotiation``'s ``"corrected"`` vs ``"review"``),
-    # rather than the raw ``supersedes_result_id is not None`` check — that
-    # naive check is wrong on a self-edit (poster corrects their own standing
-    # proposal before the recipient ever answers): it supersedes a result, but
-    # the recipient still lands on the first-post "review" view, not
-    # "corrected", so they must get the Accept/Suggest-correction prompt, not
-    # Accept/Counter.
-    is_counter = (
-        _negotiation(match, recipient_side.players[0].user_id).viewer_state
-        == "corrected"
-    )
-    copy = _result_confirmation_copy(match, poster_id, is_counter=is_counter)
-    if copy is None:
-        return
-    title, body = copy
-    result = standing_result(match)
-    push_data = {"match_id": str(match.id)}
-    if result is not None:
-        push_data["result_id"] = str(result.id)
-    for player in recipient_side.players:
-        notifications.enqueue_notification(
-            NotificationJob(
-                user_id=player.user_id,
-                category=NotificationCategory.RESULT_CONFIRM,
-                title=title,
-                body=body,
-                link=f"/matches/{match.id}",
-                action_label="Review",
-                push_category=MATCH_RESULT_CONFIRMATION_CATEGORY,
-                push_data=push_data,
-                collapse_id=f"result-confirm:{match.id}",
-            )
-        )
 
 
 # ----- scoring endpoints ---------------------------------------------------
@@ -823,7 +660,7 @@ async def create_game_score(
     except _SCORE_WRITE_ERRORS as exc:
         raise _map_score_write_error(exc) from exc
 
-    extras = await _view_extras(match_service, reloaded)
+    extras = await view_extras(match_service, reloaded)
     return _serialize_details(reloaded, current_user.id, extras)
 
 
@@ -858,7 +695,7 @@ async def update_game_score(
     except _SCORE_WRITE_ERRORS as exc:
         raise _map_score_write_error(exc) from exc
 
-    extras = await _view_extras(match_service, reloaded)
+    extras = await view_extras(match_service, reloaded)
     return _serialize_details(reloaded, current_user.id, extras)
 
 
@@ -889,7 +726,7 @@ async def delete_game_score(
     except _SCORE_WRITE_ERRORS as exc:
         raise _map_score_write_error(exc) from exc
 
-    extras = await _view_extras(match_service, reloaded)
+    extras = await view_extras(match_service, reloaded)
     return _serialize_details(reloaded, current_user.id, extras)
 
 
@@ -982,7 +819,7 @@ async def post_match_result(
         raise HTTPException(status_code=404, detail="Match not found.") from exc
 
     reloaded = outcome.match
-    extras = await _view_extras(match_service, reloaded)
+    extras = await view_extras(match_service, reloaded)
     details = _serialize_details(reloaded, current_user.id, extras)
     # Record + notify the side that now owes an acceptance. Built after the
     # response and best-effort: the result is already committed, so *nothing*
@@ -993,7 +830,7 @@ async def post_match_result(
     # clean even when the failure was the in-app persist commit.
     if outcome.awaiting_acceptance:
         try:
-            await _notify_result_posted(notifications, reloaded, current_user.id)
+            await notify_result_posted(notifications, reloaded, current_user.id)
         except Exception:
             await db.rollback()
             log.exception(
@@ -1058,5 +895,5 @@ async def accept_match_result(
             status_code=409, detail="The posted games no longer decide this match."
         ) from exc
 
-    extras = await _view_extras(match_service, reloaded)
+    extras = await view_extras(match_service, reloaded)
     return _serialize_details(reloaded, current_user.id, extras)
