+++ api/app/result_acceptance.py
@@ -36,6 +36,22 @@ from sqlalchemy.ext.asyncio import AsyncSession
 if TYPE_CHECKING:
     from app.match_scoring import _MatchWriteLoader
 
+from app.match_errors import (
+    CannotAcceptOwnProposalError,
+    MatchClosedError,
+    MatchNotFoundError,
+    MatchNotScorableError,
+    NegotiationConflictError,
+    OpponentNotFoundError,
+    PostedGamesNotDecisiveError,
+    RatedNeedsRegisteredOpponentError,
+    ResultNotFoundError,
+    ScoreConflictError,
+    ScoreNotAllowedError,
+    SelfMatchError,
+    StandingResultConflictError,
+    UndecidedBoardError,
+)
 from app.models import (
     Match,
     MatchGameScore,
@@ -55,160 +71,32 @@ from app.ratings import (
     validate_state,
 )
 from app.result_chain import standing_result
-from app.schemas.match import MatchDetailsScore
 from app.tournament_advancement import on_match_completed
 
-
-class SelfMatchError(Exception):
-    """Raised by the match-creation service when the requested opponent is the
-    acting user themselves. The HTTP adapter maps this to the existing 422
-    ``"You cannot start a match against yourself."``."""
-
-
-class OpponentNotFoundError(Exception):
-    """Raised by the match-creation service when the requested opponent id does
-    not resolve to a live (non-tombstoned) user. The HTTP adapter maps this to
-    the existing 404 ``"Opponent not found."``."""
-
-
-class RatedNeedsRegisteredOpponentError(Exception):
-    """Raised by the match-creation service when a rated match is requested
-    without a registered opponent (a solo match cannot be rated). The HTTP
-    adapter maps this to the existing 422
-    ``"A rated match needs a registered opponent."``."""
-
-
-class MatchClosedError(Exception):
-    """Raised by :func:`app.result_proposal.propose_result` when the match has
-    reached a terminal status (``completed``/``voided``) and is closed to new
-    proposals. The HTTP adapter maps this to the existing 409
-    ``"This match is no longer open to results."``. Never an ``HTTPException`` —
-    the caller adapts it to its transport."""
-
-
-class UndecidedBoardError(Exception):
-    """Raised by :func:`app.result_proposal.propose_result` when the proposed
-    board fails the strict finalize validator (empty, gappy-past-decider,
-    duplicate/out-of-range game numbers, or still-undecided) — i.e. it can't be
-    a result. Carries the validator's human-readable ``ValueError`` message; the
-    HTTP adapter maps it to the existing 422 ``str`` body. Never an
-    ``HTTPException``."""
-
-
-class NegotiationConflictError(Exception):
-    """Raised by :func:`app.result_proposal.propose_result` when the propose lost
-    the negotiation race: a first post found a result already exists, a counter
-    targeted a ``supersedes_result_id`` that is no longer the live standing
-    proposal, or the ``uq_match_results_supersedes_result_id`` unique constraint
-    tripped at commit (two concurrent counters superseding the same parent).
-
-    Carries the loaded (or reloaded) :class:`Match` so the HTTP adapter can build
-    the exact viewer-relative negotiation snapshot (``_negotiation_conflict`` /
-    ``_negotiation``) it produced before, letting a client that lost the race
-    re-render from the 409 body without an extra round-trip. Never an
-    ``HTTPException`` — the caller adapts it to its transport."""
-
-    def __init__(self, match: Match) -> None:
-        super().__init__("The standing proposal has moved on.")
-        self.match = match
-
-
-class StandingResultConflictError(Exception):
-    """Raised when the ``result_id`` handed to :func:`accept_standing_result` is
-    no longer the live standing proposal — a concurrent counter superseded it,
-    it was already accepted, or there is no standing proposal at all. The router
-    maps this to the existing 409 carrying the moved-on negotiation state."""
-
-
-class PostedGamesNotDecisiveError(Exception):
-    """Raised when the match's committed games no longer decide a winner at the
-    moment of acceptance. Practically unreachable (a result only stands once its
-    board is decided, and the board is frozen while it stands), but the core
-    stays total rather than silently stamping no winner. The router maps this to
-    the existing 409 ``"The posted games no longer decide this match."``."""
-
-
-class ResultNotFoundError(Exception):
-    """Raised by :func:`accept_result` when the target ``result_id`` is not a
-    result on the loaded match at all (as opposed to being present but no longer
-    the live standing proposal, which is :class:`StandingResultConflictError`).
-    The HTTP adapter maps this to the existing 404 ``"Result not found."``. Never
-    an ``HTTPException`` — the caller adapts it to its transport."""
-
-    def __init__(self) -> None:
-        super().__init__("Result not found.")
-
-
-class CannotAcceptOwnProposalError(Exception):
-    """Raised by :func:`accept_result` when the accepting user is a participant on
-    the *submitter's* side of the standing proposal. The proposing side already
-    consented by proposing, so only a participant on the opposing side may accept
-    (in singles, the submitter themselves can't accept their own proposal). The
-    HTTP adapter maps this to the existing 409 ``"You can't accept your own
-    proposal."``. Never an ``HTTPException`` — the caller adapts it to its
-    transport."""
-
-    def __init__(self) -> None:
-        super().__init__("You can't accept your own proposal.")
-
-
-class MatchNotFoundError(Exception):
-    """Raised by the per-game score service (``app.match_scoring``) when the
-    write target can't be resolved — either the match id is absent or the acting
-    user isn't a participant (today's score endpoints collapse both into one
-    opaque 404 so a non-participant can't probe match existence), or, on the
-    update/delete paths, the addressed game score doesn't exist.
-
-    Carries the exact ``message`` the HTTP adapter must reproduce as its 404
-    ``detail``: ``"Match not found."`` for an absent match or non-participant,
-    ``"Score not found."`` for a missing game score. Never an ``HTTPException`` —
-    it has no HTTP context; the caller adapts it to its transport."""
-
-    def __init__(self, message: str = "Match not found.") -> None:
-        super().__init__(message)
-        self.message = message
-
-
-class MatchNotScorableError(Exception):
-    """Raised by the per-game score service when the loaded match can't be
-    scored. It carries the exact ``http_status`` (422 or 409) **and** ``message``
-    for each of ``_enforce_scorable``'s four reason-specific outcomes — no
-    opponent (422), a posted result (409), an uncalled scheduled match (409), or
-    any other non-scorable/terminal state (409) — so the HTTP adapter reproduces
-    the identical response byte-for-byte while the MCP adapter reads the message.
-    Never an ``HTTPException`` — the caller adapts it to its transport."""
-
-    def __init__(self, *, http_status: int, message: str) -> None:
-        super().__init__(message)
-        self.http_status = http_status
-        self.message = message
-
-
-class ScoreNotAllowedError(Exception):
-    """Raised by the per-game score service when a scratchpad write is legal to
-    attempt but disallowed by a cross-game rule — the addressed game exceeds the
-    match's ``best_of`` range, or the prospective board would leave the match
-    decided before its last scored game (overrun). Both map to a 422 carrying
-    this exception's exact message. Never an ``HTTPException``."""
-
-
-class ScoreConflictError(Exception):
-    """Raised by the per-game score service (``app.match_scoring``) when a
-    scratchpad write loses a concurrent-participant race: a create finds the
-    game already scored (or trips the unique constraint at commit), or an
-    update's ``expected_version`` no longer matches the committed row.
-
-    Carries ``committed_score`` — the game's score as it actually stands now,
-    including its ``version`` — so the HTTP adapter can rebuild the exact 409
-    ``MatchGameScoreConflict`` body (``_score_conflict``) and a future MCP
-    adapter can point the agent back at ``get_match``. It is ``None`` only when
-    the committed row could not be re-read (the match vanished). Never an
-    ``HTTPException`` — it has no HTTP context; the caller adapts it to its
-    transport."""
-
-    def __init__(self, *, committed_score: MatchDetailsScore | None) -> None:
-        super().__init__("This game was saved by someone else while you were editing.")
-        self.committed_score = committed_score
+# The match-flow domain exception family lives in ``app.match_errors`` (a neutral
+# leaf) so the services can share it without importing one another. Re-exported
+# here because the module's own body raises several of them and because the
+# service-unit tests import them from this module.
+__all__ = [
+    "CannotAcceptOwnProposalError",
+    "MatchClosedError",
+    "MatchNotFoundError",
+    "MatchNotScorableError",
+    "NegotiationConflictError",
+    "OpponentNotFoundError",
+    "PostedGamesNotDecisiveError",
+    "RatedNeedsRegisteredOpponentError",
+    "ResultNotFoundError",
+    "ScoreConflictError",
+    "ScoreNotAllowedError",
+    "SelfMatchError",
+    "StandingResultConflictError",
+    "UndecidedBoardError",
+    "accept_result",
+    "accept_standing_result",
+    "finalize_match",
+    "side_win_counts",
+]
 
 
 def _games_to_win(best_of: int) -> int:
@@ -577,10 +465,12 @@ async def accept_result(
     monkeypatchable, ``HTTPException``-mapping load seam (the #835 row-lock race
     test barriers on it) stays the one it exercises. Never raises
     ``HTTPException`` — the caller adapts it to its transport."""
-    # Lazy imports: ``app.match_scoring`` and ``app.result_proposal`` both import
-    # this module at their top, and ``app.match_queries`` does too, so
-    # module-level imports of their symbols here would be cycles. These are only
-    # needed at call time, so this leaf stays importable ahead of them.
+    # Lazy imports: ``app.match_scoring``, ``app.result_proposal`` and
+    # ``app.match_queries`` all import this module at their top (``match_scoring``
+    # and ``match_queries`` for the scoring primitives ``_games_to_win`` /
+    # ``side_win_counts`` defined here), so module-level imports of their symbols
+    # here would be cycles. These are only needed at call time, so this leaf
+    # stays importable ahead of them.
     from app.match_queries import my_side
     from app.match_scoring import load_match_for_write
     from app.result_proposal import match_rating_eager_options
